% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eir.R
\name{calculate_eir}
\alias{calculate_eir}
\title{Calculate EIR}
\usage{
calculate_eir(m, a, Iv)
}
\arguments{
\item{m}{Ratio of vectors to humans}

\item{a}{Biting rate (bites on humans per mosquito per day)}

\item{Iv}{Proportion of infectious mosquito population}
}
\value{
EIR The Entomological Inoculation Rate
}
\description{
This function calculates the Entomological Inoculation Rate (EIR) based on the given parameters.
}
\examples{
EIR <- calculate_eir(m = 0.7, a = 0.3, Iv = 0.05)
print(EIR)
}
