% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create Malaria Model}
\usage{
create_model(params)
}
\arguments{
\item{params}{A list of parameters for the model including:
\itemize{
\item S0: Initial proportion of susceptible human population
\item Ds0: Initial proportion of diseased human population (sensitive strain)
\item As0: Initial proportion of asymptomatic human population (sensitive strain)
\item Ts0: Initial proportion of treated human population (sensitive strain)
\item DR0: Initial proportion of diseased human population (resistant strain)
\item AR0: Initial proportion of asymptomatic human population (resistant strain)
\item TR0: Initial proportion of treated human population (resistant strain)
\item Sv0: Initial proportion of susceptible mosquito population
\item Ev_s0: Initial proportion of exposed mosquito population (sensitive strain)
\item Iv_s0: Initial proportion of infectious mosquito population (sensitive strain)
\item Ev_r0: Initial proportion of exposed mosquito population (resistant strain)
\item Iv_r0: Initial proportion of infectious mosquito population (resistant strain)
\item mu: Mosquito death rate
\item e: Mosquito emergence rate
\item n: Extrinsic incubation period
\item a: Biting rate (bites on humans per mosquito per day) (optional)
\item b_h: Probability of transmission from vectors to humans
\item m: Ratio of vectors to humans (optional)
\item Phi: Proportion of symptomatic infections
\item fT: Proportion of symptomatic infections treated
\item rD: Recovery rate for diseased individuals
\item rA: Recovery rate for asymptomatic individuals
\item rTs: Recovery rate for treated individuals (sensitive strain)
\item rTR: Recovery rate for treated individuals (resistant strain)
\item Lambda_v_s: Force of infection from humans to mosquitoes (sensitive strain)
\item Lambda_v_r: Force of infection from humans to mosquitoes (resistant strain)
\item EIR_s: Entomological Inoculation Rate for sensitive strain (optional)
\item EIR_R: Entomological Inoculation Rate for resistant strain (optional)
}}
}
\value{
A list containing the \code{odin} model object, FOI (EIR_s and EIR_R), EIR_s, EIR_R, Prevalence, and Prevalence_R.
}
\description{
This function creates a malaria model using user-defined parameters and returns a list containing the \code{odin} model object, FOI, EIR_s, EIR_R, Prevalence, and Prevalence_R.
}
\examples{
params <- list(S0 = 0.9, Ds0 = 0.05, As0 = 0.025, Ts0 = 0.0125,
               DR0 = 0.005, AR0 = 0.0025, TR0 = 0.001,
               a = 0.3, b_h = 0.1, m = 0.7, Phi = 0.7,
               fT = 0.2, rD = 0.05, rA = 0.03, rTs = 0.02,
               rTR = 0.01, Sv0 = 0.9, Ev_s0 = 0, Iv_s0 = 0,
               Ev_r0 = 0, Iv_r0 = 0,
               e = 0.1, mu = 0.1, n = 10,
               Lambda_v_s = 0.005, Lambda_v_r = 0.003,
               EIR_s = NULL, EIR_R = NULL)
results <- create_model(params)
model_results <- results$model$run(0:500)
prevalence <- rowSums(model_results[, c("AR", "TR", "DR", "As", "Ds", "Ts")])
print(prevalence)
print(results$FOI)
print(results$EIR_s)
print(results$EIR_R)
}
