% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_malaria_model.R
\name{plot_malaria_model}
\alias{plot_malaria_model}
\title{Plot Malaria Model Results}
\usage{
plot_malaria_model(
  results,
  colors = NULL,
  line_width = 1,
  legend_position = "right"
)
}
\arguments{
\item{results}{The output from the \code{run} method of the malaria model.}

\item{colors}{A vector of colors for the population lines. If NULL, default colors will be used.}

\item{line_width}{The width of the population lines.}

\item{legend_position}{The position of the legend ("top", "bottom", "left", "right", or "none").}
}
\description{
This function takes the output of the malaria model and creates a plot showing the dynamics of each population over time.
}
