% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot Malaria Model Results}
\usage{
plot_model(
  model_results,
  res_time,
  ton,
  toff = NULL,
  output_vars = NULL,
  colors = NULL,
  line_types = NULL,
  line_widths = NULL,
  title = "Malaria Model Results",
  subtitle = NULL,
  show_lines = c("res_time", "ton", "toff"),
  vline_colors = NULL,
  vline_types = NULL,
  vline_widths = NULL
)
}
\arguments{
\item{model_results}{A data frame containing the model results or an odin model object.}

\item{res_time}{The time at which resistance was introduced.}

\item{ton}{The time at which treatment was turned on.}

\item{toff}{The time at which treatment was turned off.}

\item{output_vars}{A character vector specifying the output variables to plot. If NULL, all output variables will be plotted.}

\item{colors}{A named character vector specifying the colors for each variable. If NULL, default colors will be used.}

\item{line_types}{A named character vector specifying the line types for each variable. If NULL, default line types will be used.}

\item{line_widths}{A named numeric vector specifying the line widths for each variable. If NULL, default line widths will be used.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot. If NULL, a default subtitle will be generated.}

\item{show_lines}{A character vector specifying which vertical lines to show. Can include "res_time", "ton", and "toff".}

\item{vline_colors}{A named character vector specifying the colors for each vertical line. Default colors are provided.}

\item{vline_types}{A named character vector specifying the line types for each vertical line. Default line types are provided.}

\item{vline_widths}{A named numeric vector specifying the line widths for each vertical line. Default widths are provided.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of a malaria model simulation.
}
