% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range.R
\name{plot_range}
\alias{plot_range}
\title{Plot Range Model Results}
\usage{
plot_range(
  model_results,
  x_var,
  y_var,
  color_var,
  output_vars = NULL,
  colors = NULL,
  line_types = NULL,
  line_widths = NULL,
  title = "Range Model Results",
  subtitle = NULL
)
}
\arguments{
\item{model_results}{A data frame containing the model results from \code{range_model}.}

\item{x_var}{A string specifying the variable to be used for the x-axis.}

\item{y_var}{A string specifying the variable to be used for the y-axis.}

\item{color_var}{A string specifying the variable to be used for color classification.}

\item{output_vars}{A character vector specifying the output variables to plot. If NULL, all output variables will be plotted.}

\item{colors}{A named character vector specifying the colors for each variable. If NULL, default colors will be used.}

\item{line_types}{A named character vector specifying the line types for each variable. If NULL, default line types will be used.}

\item{line_widths}{A named numeric vector specifying the line widths for each variable. If NULL, default line widths will be used.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot.}
}
\value{
A ggplot object.
}
\description{
Plot Range Model Results
}
