% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_model.R
\name{range_model}
\alias{range_model}
\title{Run Malaria Model for Ranges of Parameters}
\usage{
range_model(
  param_ranges,
  params = NULL,
  output_vars = NULL,
  use_eir_ft = FALSE,
  run_time = 5000,
  time_points = NULL,
  time_scale = "day",
  ton = 5000,
  toff = 50000,
  init_res = 0.01,
  res_time = 3000,
  rTR_true = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{param_ranges}{A named list of parameter ranges. Each element should be a vector of values to run the model over.}

\item{params}{A list of other parameters to pass to the \code{malaria_model} function}

\item{output_vars}{A character vector specifying the output variables to return. If NULL, all output variables will be returned.}

\item{use_eir_ft}{Logical. If TRUE, uses EIR and ft to generate parameters. If FALSE, uses provided params.}

\item{run_time}{Numeric. The total time to run the model. Default is 5000.}

\item{time_points}{Numeric vector. Specific time points to output results. If NULL, all time points will be returned.}

\item{time_scale}{Character. Either "day" or "year". Default is "day". This affects both input and output time units.}

\item{ton}{Time at which treatment is turned on}

\item{toff}{Time at which treatment is turned off}

\item{init_res}{Initial resistance level}

\item{res_time}{Time at which resistance is introduced}

\item{rTR_true}{True treatment rate for resistant parasites}

\item{verbose}{Logical. If TRUE, prints detailed logs. Default is FALSE.}
}
\value{
A data frame of results for each combination of parameter values
}
\description{
Run Malaria Model for Ranges of Parameters
}
