% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_model.R
\name{summarize_model_results}
\alias{summarize_model_results}
\title{Summarize Model Results}
\usage{
summarize_model_results(results, time_points)
}
\arguments{
\item{results}{A data frame containing the model results.}

\item{time_points}{A vector of specific time points to summarize results.}
}
\value{
A summarized data frame.
}
\description{
Summarize Model Results
}
