% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_model.R
\name{form_pars}
\alias{form_pars}
\title{Format parameters for the malaria model}
\usage{
form_pars(pars)
}
\arguments{
\item{pars}{Parameter List}
}
\value{
Formatted parameter list
}
\description{
Format parameters for the malaria model
}
