% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot Malaria Model Results}
\usage{
plot_model(
  model_results,
  res_time,
  ton,
  toff = NULL,
  output_vars = NULL,
  colors = NULL,
  line_types = NULL,
  line_widths = NULL,
  title = "Malaria Model Results",
  subtitle = NULL,
  show_lines = c("res_time", "ton", "toff"),
  vline_colors = NULL,
  vline_types = NULL,
  vline_widths = NULL
)
}
\arguments{
\item{model_results}{A data frame containing the model results.}

\item{res_time}{The time at which resistance was introduced.}

\item{ton}{The time at which treatment was turned on.}

\item{toff}{The time at which treatment was turned off.}

\item{output_vars}{A character vector specifying the output variables to plot. If NULL, all output variables will be plotted.}

\item{colors}{A named character vector specifying the colors for each variable. If NULL, default colors will be used.}

\item{line_types}{A named character vector specifying the line types for each variable. If NULL, default line types will be used.}

\item{line_widths}{A named numeric vector specifying the line widths for each variable. If NULL, default line widths will be used.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot. If NULL, a default subtitle will be generated.}

\item{show_lines}{A character vector specifying which vertical lines to show. Can include "res_time", "ton", and "toff".}

\item{vline_colors}{A named character vector specifying the colors for each vertical line. Default colors are provided.}

\item{vline_types}{A named character vector specifying the line types for each vertical line. Default line types are provided.}

\item{vline_widths}{A named numeric vector specifying the line widths for each vertical line. Default widths are provided.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of a malaria model simulation.
}
\examples{
\donttest{
library(AMRSpreadModel)

params <- list(
  S0 = 0.9, Ds0 = 0.05, As0 = 0.02, Ts0 = 0.03, DR0 = 0, AR0 = 0,
  TR0 = 0, Sv0 = 0.9, Ev_s0 = 0.08, Iv_s0 = 0.02, Ev_r0 = 0, Iv_r0 = 0,
  m = 10, a = 0.3, b = 0.5876259, phi = 0.7, fT = 0.1, rD = 0.2,
  rA = 0.01, rTs = 0.2, rTR_true = 0.01, e = 0.132, mu = 0.132,
  n = 10, c_A = 0.05, c_D = 0.06, c_T = 0.02, ton = 8000,
  toff = 99999999, res_time = 3000, init_res = 0.2
)

model <- malaria_model(params = params)
if (!is.null(model)) {
  results <- model$run(0:500)
  plot_model(results, res_time = params$res_time, ton = params$ton, toff = params$toff)
} else {
  print("Failed to create model. Check parameters and try again.")
}
}
}
